Fix: delbtn CSS
