+++ src/components/Post/PostContainer.jsx
@@ -17,6 +17,7 @@ const PostContainer = () => {
   const [content, setContent] = useState("");
   const [img, setImg] = useState("");
 
+  // input value > state
   const typeInfo = (type) => {
     setType(type);
   };
@@ -30,8 +31,6 @@ const PostContainer = () => {
     setImg(img);
   };
 
-  // BE요청 후 수정 / 추가
-
   const addPost = {
     style: style,
     type: type,
@@ -39,23 +38,16 @@ const PostContainer = () => {
   };
 
   const addClickHandler = () => {
-    // console.log("addPost", addPost);
-    console.log("addPost", addPost);
-
     let formData = new FormData();
     formData.append(
       "data",
       new Blob([JSON.stringify(addPost)], { type: "application/json" })
     );
-
     formData.append("file", img);
-
     for (let i of formData.entries()) {
       console.log("i", i[1]);
     }
-    // console.log("formData", formData);
     dispatch(_addPost(formData));
-
     navigate("/community");
   };
